requirements.txt
discord.py==2.3.2
fastapi==0.115.4
uvicorn[standard]==0.32.0
requests==2.32.3
beautifulsoup4==4.12.3
